+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/command_generators/command_generator_base.py
@@ -12,6 +12,7 @@ methods.
 
 from __future__ import annotations
 
+import inspect
 import torch
 import weakref
 from abc import ABC, abstractmethod
@@ -48,6 +49,7 @@ class CommandGeneratorBase(ABC):
         # store the inputs
         self.cfg = cfg
         self._env = env
+
         # create buffers to store the command
         # -- metrics that can be used for logging
         self.metrics = dict()
@@ -56,21 +58,16 @@ class CommandGeneratorBase(ABC):
         # -- counter for the number of times the command has been resampled within the current episode
         self.command_counter = torch.zeros(self.num_envs, device=self.device, dtype=torch.long)
 
-        # add callback for debug visualization
-        if self.cfg.debug_vis:
-            app_interface = omni.kit.app.get_app_interface()
-            # NOTE: Use weakref on callback to ensure that this object can be deleted when its destructor is called.
-            self._debug_visualization_handle = app_interface.get_post_update_event_stream().create_subscription_to_pop(
-                lambda event, obj=weakref.proxy(self): obj._debug_vis_callback(event),
-            )
-        else:
-            self._debug_visualization_handle = None
+        # add handle for debug visualization (this is set to a valid handle inside set_debug_vis)
+        self._debug_vis_handle = None
+        # set initial state of debug visualization
+        self.set_debug_vis(self.cfg.debug_vis)
 
     def __del__(self):
         """Unsubscribe from the callbacks."""
-        if self._debug_visualization_handle is not None:
-            self._debug_visualization_handle.unsubscribe()
-            self._debug_visualization_handle = None
+        if self._debug_vis_handle:
+            self._debug_vis_handle.unsubscribe()
+            self._debug_vis_handle = None
 
     """
     Properties
@@ -92,21 +89,47 @@ class CommandGeneratorBase(ABC):
         """The command tensor. Shape is (num_envs, command_dim)."""
         raise NotImplementedError
 
+    @property
+    def has_debug_vis_implementation(self) -> bool:
+        """Whether the command generator has a debug visualization implemented."""
+        # check if function raises NotImplementedError
+        source_code = inspect.getsource(self._debug_vis_callback)
+        return "NotImplementedError" not in source_code
+
     """
     Operations.
     """
 
-    def set_debug_vis(self, debug_vis: bool):
+    def set_debug_vis(self, debug_vis: bool) -> bool:
         """Sets whether to visualize the command data.
 
         Args:
             debug_vis: Whether to visualize the command data.
 
-        Raises:
-            RuntimeError: If the command debug visualization is not enabled.
+        Returns:
+            Whether the debug visualization was successfully set. False if the command
+            generator does not support debug visualization.
         """
-        if not self.cfg.debug_vis:
-            raise RuntimeError("Debug visualization is not enabled for this sensor.")
+        # check if debug visualization is supported
+        if not self.has_debug_vis_implementation:
+            return False
+        # toggle debug visualization objects
+        self._set_debug_vis_impl(debug_vis)
+        # toggle debug visualization handles
+        if debug_vis:
+            # create a subscriber for the post update event if it doesn't exist
+            if self._debug_vis_handle is None:
+                app_interface = omni.kit.app.get_app_interface()
+                self._debug_vis_handle = app_interface.get_post_update_event_stream().create_subscription_to_pop(
+                    lambda event, obj=weakref.proxy(self): obj._debug_vis_callback(event)
+                )
+        else:
+            # remove the subscriber if it exists
+            if self._debug_vis_handle is not None:
+                self._debug_vis_handle.unsubscribe()
+                self._debug_vis_handle = None
+        # return success
+        return True
 
     def reset(self, env_ids: Sequence[int] | None = None) -> dict[str, float]:
         """Reset the command generator and log metrics.
@@ -174,14 +197,6 @@ class CommandGeneratorBase(ABC):
         self._resample_command(env_ids)
 
     """
-    Simulation callbacks.
-    """
-
-    def _debug_vis_callback(self, event):
-        """Visualizes the sensor data."""
-        self._debug_vis_impl()
-
-    """
     Implementation specific functions.
     """
 
@@ -200,9 +215,18 @@ class CommandGeneratorBase(ABC):
         """Update the metrics based on the current state."""
         raise NotImplementedError
 
-    def _debug_vis_impl(self):
-        """Visualize the command in the simulator.
+    def _set_debug_vis_impl(self, debug_vis: bool):
+        """Set debug visualization into visualization objects.
 
-        This is an optional function that can be used to visualize the command in the simulator.
+        This function is responsible for creating the visualization objects if they don't exist
+        and input ``debug_vis`` is True. If the visualization objects exist, the function should
+        set their visibility into the stage.
         """
-        pass
+        raise NotImplementedError(f"Debug visualization is not implemented for {self.__class__.__name__}.")
+
+    def _debug_vis_callback(self, event):
+        """Callback for debug visualization.
+
+        This function calls the visualization objects and sets the data to visualize into them.
+        """
+        raise NotImplementedError(f"Debug visualization is not implemented for {self.__class__.__name__}.")